requirements.txt
# Dependencias principales
python-dotenv

# Proveedores de IA (instala solo los que necesites)
openai                    # Para OpenAI y Azure OpenAI
google-genai      # Para Google Gemini

# Dependencias de desarrollo (opcional)
# pytest>=7.0.0                 # Para testing
# black>=23.0.0                 # Para formateo de código
# flake8>=6.0.0                 # Para linting